refactor(alignment): streamline audio handling by consolidating temporary WAV file creation and updating alignment model interface
